pyproject.toml
[project]
name = "whatsnew"
version = "0.1.0"
description = "An AI Agent that tells you what's new/trending in a topic of your choosing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["langchain", "langgraph", "python-dotenv", "langchain-anthropic", "pyyaml", "pydantic"]
